% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{brms_model}
\alias{brms_model}
\title{Interface for model construction}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{trending}{\code{\link[trending:trending_model]{brms_model}}}
}}

