% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect-changepoint.R
\name{detect_changepoint}
\alias{detect_changepoint}
\title{Detect change in temporal trend}
\usage{
detect_changepoint(
  data,
  models,
  alpha = 0.05,
  max_k = 7,
  method = trending::evaluate_resampling,
  ...
)
}
\arguments{
\item{data}{A \code{data.frame} or a \code{tibble} containing the response and
explanatory variables used in the \code{models}.}

\item{models}{A list of \url{trendbreaker_model} objects,
generated by \code{lm_model}, \code{glm_model}, \code{glm_nb_model}, \code{brms_model} and
similar functions (see \code{?trendbreaker_model}) for details.}

\item{alpha}{The alpha threshold to be used for the prediction interval
calculation; defaults to 0.05, i.e. 95\% prediction intervals are
calculated.}

\item{max_k}{An \code{integer} indicating the maximum number of recent data points
to be excluded from the trend fitting procedure. By default, ASMODEE will
look for a changepoint within this recent time period, after which data no
longer fit the previous trend. Larger values will require more computation
from the method. Only used if \code{fixed_k} is \code{NULL}.}

\item{method}{A function used to evaluate model fit. Current choices are
\code{evaluate_resampling} (default) and \code{evaluate_aic}. \code{evaluate_resampling}
uses cross-validation and RMSE to assess model fit. \code{evaluate_aic} uses
Akaike's Information Criterion instead, which is faster but possibly less
good a selecting models with the best predictive power.}

\item{...}{Further arguments passed to \code{method}.}
}
\description{
This function looks for the 'optimal' number of recent days to exclude from
the temporal trend fitting in \url{asmodee}. The procedure selects
the value of \code{k} which maximises the sum if i) the number of non-outliers in
training set (before the last \code{k} points) ii) the number of outliers in the
last \code{k} points. Note that for each value of \code{k} investigated, model
selection is performed as described in \code{?asmodee}.
}
\seealso{
\url{asmodee}
}
\author{
Thibaut Jombart
}
