% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect-outliers.R
\name{detect_outliers}
\alias{detect_outliers}
\title{Detect outliers given a fitted model}
\usage{
detect_outliers(data, model, alpha = 0.05, ...)
}
\arguments{
\item{data}{a \code{data.frame} containing data for which predictions are to be
derived}

\item{model}{a fitted model as \code{\link[trending:fit]{trending::trending_model_fit}} object;
this can be obtained by running \code{\link[trending:fit]{trending::fit()}} on a
\code{\link[trending:trending_model]{trending::trending_model}}.}

\item{alpha}{the alpha threshold to be used for prediction intervals,
defaulting to 0.05, i.e. 95\% prediction intervals are derived}

\item{...}{additional arguments passed to the underlying model prediction
function}
}
\value{
A \code{data.frame} containing the original data, the mean model
predictions, the lower and upper bounds of the prediction interval, a
variable \code{outlier} indicating which point is an outlier as a \code{logical}, and
the \code{classification} of data points as a \code{factor} with values \code{normal} for
points within the PI, \code{increase} for outliers above the PI, and \code{decrease}
for outliers below the PI.
}
\description{
This function uses a trained model stored as an\code{trendbing_model_fit} object
to derive the prediction interval (PI) for a given alpha threshold; it
classifies as outliers every data point falling outside the interval.
}
\author{
Thibaut Jombart, Dirk Schumacher
}
