% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epichange-accessors.R, R/model-accessors.R
\name{get_model.epichange}
\alias{get_model.epichange}
\alias{epichange-accessors}
\alias{epichange-class}
\alias{get_k.epichange}
\alias{get_results.epichange}
\alias{get_outliers.epichange}
\alias{predict.epichange}
\alias{get_formula}
\title{Accessors for epichange objects}
\usage{
\method{get_model}{epichange}(x)

\method{get_k}{epichange}(x)

\method{get_results}{epichange}(x)

\method{get_outliers}{epichange}(x)

\method{predict}{epichange}(object, newdata, alpha = 0.05, ...)

get_formula(model)
}
\arguments{
\item{model}{the output of functions \code{lm_model}, \code{glm_model}, or
\code{glm_nb_model}}
}
\description{
These functions can be used to access information stored in \code{epichange}
objects. See details.

These functions can be used to access information stored in \code{epichange}
objects. See details.
}
\details{
The following accessors are available:
\itemize{
\item \code{get_model}: fitted model capturing the temporal trend in the data as an
\code{epichange_model_fit} object
\item \code{get_k}: number of recent days excluded from the temporal trend
\item \code{get_results}: main \code{data.frame} containing the original data, the
predicted values with lower and upper bounds for the prediction interval, a
\code{logical} variable \code{outlier} which is \code{TRUE} for data points falling outside
the prediction interval, and \code{classification} of outliers as a \code{factor}.
\item \code{get_outliers}: returns data points (as rows of \code{get_results(x)}
corresponding to outliers
\item \code{get_classification}: returns a \code{factor} indicating if data points are
\code{normal} outliers (\code{TRUE}) or not (\code{FALSE})
}

The following accessors are available:
\itemize{
\item \code{get_model}: fitted model capturing the temporal trend in the data as an
\code{epichange_model_fit} object
\item \code{get_k}: number of recent days excluded from the temporal trend
\item \code{get_results}: main \code{data.frame} containing the original data, the
predicted values with lower and upper bounds for the prediction interval, a
\code{logical} variable \code{outlier} which is \code{TRUE} for data points falling outside
the prediction interval, and \code{classification} of outliers as a \code{factor}.
\item \code{get_outliers}: returns data points (as rows of \code{get_results(x)}
corresponding to outliers
\item \code{get_classification}: returns a \code{factor} indicating if data points are
\code{normal} outliers (\code{TRUE}) or not (\code{FALSE})
}
}
\author{
Thibaut Jombart, Dirk Schumacher

Thibaut Jombart, Dirk Schumacher
}
