% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_training_data.R
\name{get_training_data}
\alias{get_training_data}
\title{Get training data}
\usage{
get_training_data(x, date_index, k)
}
\arguments{
\item{x}{A \code{data.frame} or a \code{tibble}.}

\item{date_index}{The name of the variable in \code{df} to be used as time.}

\item{k}{The number of the last data points to exclude from the training
set. These will be the last \code{k} unique values of the \code{time} column.}
}
\description{
This function removes the last 'k' time units from a dataset to define a
training set. Time is known through the name of a column indicating time.
}
