% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.epichange}
\alias{plot.epichange}
\title{Plotting method for epichange objects}
\usage{
\method{plot}{epichange}(
  x,
  x_axis,
  point_size = 2,
  col_normal = "#8B8B8C",
  col_increase = "#CB3355",
  col_decrease = "#32AB96",
  guide = TRUE,
  ...
)
}
\arguments{
\item{x}{an \code{epichange} object, as returned by \code{asmodee}}

\item{x_axis}{the name or position of the variable in \code{get_results(x)} to be
used on the x-axis, which represents time}

\item{point_size}{the size of the points to be used; defaults to 2}

\item{col_normal}{the color to be used for non-outlying observations,
i.e. observations falling within the prediction interval of the estimated
temporal trend}

\item{col_increase}{the color to be used for outlying observations which are
above the prediction interval of the estimated temporal trend}

\item{col_decrease}{the color to be used for outlying observations which are
below the prediction interval of the estimated temporal trend}

\item{guide}{a \code{logical} indicating whether a color legend should be added to
the plot (\code{TRUE}, default) or not (\code{FALSE})}

\item{...}{unused - present for compatibility with the \code{plot} generic}
}
\description{
The plotting method for \code{epichange} objects produces a \code{ggplot} object, which
can then be modified using \code{ggplot2}. It accepts a few arguments for
customising the graphs produced.
}
