% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_training_data.R
\name{set_training_data}
\alias{set_training_data}
\title{Set training data}
\usage{
set_training_data(x, date_index, k)
}
\arguments{
\item{x}{A \code{data.frame} or a \code{tibble}.}

\item{date_index}{The name of the variable in \code{df} to be used as time.}

\item{k}{The number of the last data points to exclude from the training
set. These will be the last \code{k} unique values of the \code{time} column.}
}
\description{
This function adds a column 'training' to a \code{data.frame}, a \code{logical}
indicating if data points are part of the training set ('callibrartion
window'), or not. The last 'k' time units from a dataset to define a training
set. Time is known through the name of a column indicating time.
}
