% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.trendbreaker}
\alias{summary.trendbreaker}
\alias{summary.trendbreaker_incidence2}
\title{Summary for trendbreaker outputs}
\usage{
\method{summary}{trendbreaker}(object, ...)

\method{summary}{trendbreaker_incidence2}(object, ...)
}
\arguments{
\item{object}{the \code{trendbreaker} object to summarise}

\item{...}{further arguments to be passed to other functions (currently ignored)}
}
\value{
a \code{data.frame} containing the following colums:
\itemize{
\item n_recent: number of data points in recent set
\item n_recent_increases: number of recent increases
\item n_recent_decreases: number of recent decreases
\item n_recent_outliers: number of recent outliers
\item n_training: number of data points in training set
\item n_training_increases: number of increases in training set
\item n_training_decreases: number of decreases in training set
\item n_training_outliers: number of outliers in training set
}
}
\description{
Summary for trendbreaker outputs
}
\author{
Thhibaut Jombart
}
