% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trendbreaker-accessors.R
\name{get_model.trendbreaker}
\alias{get_model.trendbreaker}
\alias{trendbreaker-accessors}
\alias{trendbreaker-class}
\alias{get_k.trendbreaker}
\alias{get_results.trendbreaker}
\alias{get_outliers.trendbreaker}
\alias{predict.trendbreaker}
\title{Accessors for trendbreaker objects}
\usage{
\method{get_model}{trendbreaker}(x, ...)

\method{get_k}{trendbreaker}(x, ...)

\method{get_results}{trendbreaker}(x, ...)

\method{get_outliers}{trendbreaker}(x, ...)

\method{predict}{trendbreaker}(object, newdata, alpha = 0.05, ...)
}
\arguments{
\item{x}{an \code{trendbreaker} object, as returned by \url{asmodee}}

\item{...}{further arguments passed to other methods}

\item{object}{an \code{trendbreaker} object, as returned by \url{asmodee}}

\item{newdata}{a \code{data.frame} containing data for which predictions are to be
derived}

\item{alpha}{the alpha threshold to be used for prediction intervals,
defaulting to 0.05, i.e. 95\% prediction intervals are derived}
}
\description{
These functions can be used to access information stored in \code{trendbreaker}
objects. See details.
}
\details{
The following accessors are available:
\itemize{
\item \code{get_model()}: fitted model capturing the temporal trend in the data as an
\code{trendbreaker_model_fit} object
\item \code{get_k()}: number of recent days excluded from the temporal trend
\item \code{get_results()}: main \code{data.frame} containing the original data, the
predicted values with lower and upper bounds for the prediction interval, a
\code{logical} variable \code{outlier} which is \code{TRUE} for data points falling outside
the prediction interval, and \code{classification} of outliers as a \code{factor}.
\item \code{get_outliers()}: returns data points (as rows of \code{get_results(x)}
corresponding to outliers
\item \code{predict()}: function to make model predictions from the fitted model in the
\code{trendbreaker} object; accepts two arguments \code{newdata}, a mandatory input
containing data for which predictions are derived, and \code{alpha}, the threshold
used for prediction intervals, defaulting to 0.05.
}
}
\author{
Thibaut Jombart, Dirk Schumacher
}
