% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trendbreaker-model-accessors.R
\name{get_formula.trendbreaker_model}
\alias{get_formula.trendbreaker_model}
\alias{trendbreaker_model-accessors}
\alias{trendbreaker_model-class}
\alias{get_response.trendbreaker_model}
\alias{get_family.trendbreaker_model}
\alias{train.trendbreaker_model}
\title{Accessors for trendbreaker_model objects}
\usage{
\method{get_formula}{trendbreaker_model}(x, ...)

\method{get_response}{trendbreaker_model}(x, ...)

\method{get_family}{trendbreaker_model}(x, ...)

\method{train}{trendbreaker_model}(x, data, ...)
}
\arguments{
\item{x}{the output of functions \code{lm_model}, \code{glm_model}, or
\code{glm_nb_model}}

\item{...}{further arguments passed to other methods}

\item{data}{a \code{data.frame} to be used to train the model#'}
}
\description{
These functions can be used to access information stored in \code{trendbreaker_model}
objects. See details.
}
\details{
The following accessors are available:
\itemize{
\item \code{get_formula()}: get the formula used to model temporal trends
\item \code{get_response()}: get the name of the response variable
\item \code{get_family()}: get the model family, indicating the type of distribution
assumed for the response variable
\item \code{train()}: train a model using data to obtain a
\url{trendbreaker_model_fit} object
}
}
\author{
Thibaut Jombart, Dirk Schumacher
}
