% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epichange-model-accessors.R
\name{get_formula.epichange_model}
\alias{get_formula.epichange_model}
\alias{epichange_model-accessors}
\alias{epichange_model-class}
\alias{get_formula}
\alias{get_formula.epichange}
\alias{get_response.epichange_model}
\alias{get_response}
\alias{get_family.epichange_model}
\alias{get_family}
\alias{train.epichange_model}
\alias{train}
\title{Accessors for epichange_model objects}
\usage{
\method{get_formula}{epichange_model}(model)

\method{get_response}{epichange_model}(model)

\method{get_family}{epichange_model}(model)

\method{train}{epichange_model}(model, data)
}
\arguments{
\item{model}{the output of functions \code{lm_model}, \code{glm_model}, or
\code{glm_nb_model}}

\item{data}{a \code{data.frame} to be used to train the model#'}
}
\description{
These functions can be used to access information stored in \code{epichange_model}
objects. See details.
}
\details{
The following accessors are available:
\itemize{
\item \code{get_formula}: get the formula used to model temporal trends
\item \code{get_response}: get the name of the response variable
\item \code{get_family}: get the model family, indicating the type of distribution
assumed for the response variable
\item \code{train}: train a model using data to obtain a
\url{epichange_model_fit} object
}
}
\author{
Thibaut Jombart, Dirk Schumacher
}
