% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{evaluate_resampling}
\alias{evaluate_resampling}
\title{Model evaluation}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{trending}{\code{\link[trending:evaluate_models]{evaluate_resampling}}}
}}

