% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{get_model}
\alias{get_model}
\alias{trendbreaker-generics}
\alias{get_k}
\alias{get_results}
\alias{get_outliers}
\alias{get_formula}
\alias{get_response}
\alias{get_family}
\alias{train}
\title{S3 generics for trendbreaker}
\usage{
get_model(x, ...)

get_k(x, ...)

get_results(x, ...)

get_outliers(x, ...)

get_formula(x, ...)

get_response(x, ...)

get_family(x, ...)

train(x, data, ...)
}
\arguments{
\item{x}{the object to access information from}

\item{...}{further arguments used in methods}

\item{data}{a \code{data.frame} to be used as training set for the model}
}
\description{
These are generic functions used by the \emph{trendbreaker} package, mostly used for
accessing content of various objects. See \code{?trendbreaker-accessors} for methods
relating to \code{trendbreaker} objects, and \code{trendbreaker_model-accessors} for methods
relating to \code{trendbreaker_model} objects.
}
\seealso{
\url{trendbreaker-accessors},
\url{trendbreaker_model-accessors}
}
