% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{get_k}
\alias{get_k}
\alias{trendbreaker-generics}
\alias{get_results}
\alias{get_outliers}
\title{S3 generics for trendbreaker}
\usage{
get_k(x, ...)

get_results(x, ...)

get_outliers(x, ...)
}
\arguments{
\item{x}{a \code{data.frame} to be used as training set for the model}

\item{...}{further arguments used in methods}
}
\description{
These are generic functions used by the \emph{trendbreaker} package, mostly used for
accessing content of various objects. See \code{?trendbreaker-accessors} for
associated methods.
}
\seealso{
\url{trendbreaker-accessors},
}
