% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trendbreaker-package.R
\docType{package}
\name{trendbreaker-package}
\alias{trendbreaker}
\alias{trendbreaker-package}
\title{trendbreaker: Detect Changes in Temporal Trends}
\description{
Provides tools for detecting changes in temporal trends of a single response variable. It provides a coherent interface to several modeling tools, alongside functions for model selection and outlier detection. It implements the Automatic Selection of Models and Outlier Detection for Epidemmics (ASMODEE), an algorithm originally designed for detecting changes in COVID-19 case incidence.
}
\author{
\strong{Maintainer}: Dirk Schumacher \email{mail@dirk-schumacher.net}

Authors:
\itemize{
  \item Thibaut Jombart \email{thibautjombart@gmail.com}
}

}
\keyword{internal}
